Update generic-bigtreetech-skr-mini-e3.cfg
There were some errors to this configuration file which prevented it from working.  The endstop pins do not need to be inverted, the x/y and extruder dir pins need inverted, The step distances were all wrong as well.  Also added the stock display for Ender 3 and my PID tunings (these may not be necessary).  May not be perfect, but this at least works.